Bug on CloudEndpoints/2 - Type:String, Format:Int64
When executing the code generated from a CloudEndpoint, with a resource with a property of type Long(Java), which appears in the discovery document as type:string and format:int64, throws an exception because tries to cast a string from the json to a int.

Related to the issue #42 and #38.

Scarygami pointed that the current implementation of Cloud Endpoints for Python correctly returns a number in the JSON, so the previous solution didn't work. What about this one?